rename module containing assertion functions
